docs: update documentation for 3.0
With the 3.0.0 release the library now has completely separate scripts
respective to Vue.js versions both 2.x and 3.x. In order to import the
plugin the version descriptior of the vue version used must be appended
to the import path. (`import mdiVue from 'mdi-vue/v2` for Vue.js 2.x
or `mdi-vue/v3` for Vue.js 3.x)